feat(admin): move admin page content into slideshow component
- Extracted the admin page into a reusable AdminAnnouncementPage component.
- The content is now wrapped in Splide slides, enabling integration of additional pages within the /admin route.
